requirements.txt
fastapi
uvicorn[standard]
python-multipart
assemblyai
python-dotenv
yt-dlp
httpx
langgraph
openai
websockets
pydantic
tenacity
structlog
prometheus-client
# dev / eval
pytest
pytest-asyncio
jiwer
